+++ components/elastic_band/src/specificworker.h
@@ -73,8 +73,8 @@ class SpecificWorker : public GenericWorker
         int startup_check();
         void initialize(int period);
         void new_target_from_mouse(int pos_x, int pos_y, int id);
-        void update_node_slot(const std::int32_t id, const std::string &type);
-        void update_attrs_slot(const std::int32_t id, const std::map<string, DSR::Attribute> &attribs);
+        void update_node_slot(const std::uint64_t id, const std::string &type);
+        void update_attrs_slot(const std::uint64_t id, const std::map<string, DSR::Attribute> &attribs);
 
 private:
         // DSR graph
@@ -114,7 +114,7 @@ private:
         const float ROAD_STEP_SEPARATION = ROBOT_LENGTH * 0.9;
         float KE = 40;
         float KI = 300;
-        std::uint32_t last_path_id;  // ID of last path node that came through the slot
+        std::uint64_t last_path_id;  // ID of last path node that came through the slot
 
         enum class SearchState {NEW_TARGET, AT_TARGET, NO_TARGET_FOUND, NEW_FLOOR_TARGET};
         void elastic_band_initialize( );
