+++ components/octomap/src/specificworker.h
@@ -53,7 +53,7 @@ public slots:
 	void compute();
 	int startup_check();
 	void initialize(int period);
-    void update_node_slot(const std::int32_t id, const std::string &type);
+    void update_node_slot(const std::uint64_t id, const std::string &type);
 
 private:
 	// DSR graph
